change names of CenterOne and METH_Sensor
