+++ SC_backend/slow_control_code/CenterOne/Makefile
@@ -5,11 +5,11 @@ CFLAGS = `mariadb_config --cflags` -I../include
 
 LDFLAGS =  -L../lib -lslow_control -I/usr/local/include 
 
-PROGS = CenterThree
+PROGS = CenterOne
 
 all: $(PROGS) 
 
-HiCube80: $(PROGS).o
+CenterOne: $(PROGS).o
 	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ $@.o $(LIBS)
 	rm -f $@.o
