+++ SC_backend/slow_control_code/Makefile
@@ -26,6 +26,9 @@ SUBDIRS = lib \
 	TPG26X \
 	WIENER_VME \
 	Sentry_PDU\
+	CenterOne\
+	METH_SENSOR\
+	
 
 .PHONY: $(SUBDIRS)
